chore: rename cloud-runner-manager to vm-manager
