+++ github-runner-manager/src/github_runner_manager/manager/vm_manager.py
@@ -94,33 +94,6 @@ class CloudRunnerState(str, Enum):
         return state
 
 
-class CloudInitStatus(str, Enum):
-    """Represents the state of cloud-init script.
-
-    The cloud init script is used to launch ephemeral GitHub runners. If the script is being
-    initialized, GitHub runner is listening for jobs or GitHub runner is running the job, the
-    cloud-init script should report "running" status.
-
-    Refer to the official documentation on cloud-init status:
-    https://cloudinit.readthedocs.io/en/latest/howto/status.html.
-
-    Attributes:
-        NOT_STARTED: The cloud-init script has not yet been started.
-        RUNNING: The cloud-init script is running.
-        DONE: The cloud-init script has completed successfully.
-        ERROR: There was an error while running the cloud-init script.
-        DEGRADED: There was a non-critical issue while running the cloud-inits script.
-        DISABLED: Cloud init was disabled by other system configurations.
-    """
-
-    NOT_STARTED = "not started"
-    RUNNING = "running"
-    DONE = "done"
-    ERROR = "error"
-    DEGRADED = "degraded"
-    DISABLED = "disabled"
-
-
 @dataclass
 class CloudRunnerInstance:
     """Information on the runner on the cloud.
@@ -261,6 +234,7 @@ class CloudRunnerManager(abc.ABC):
         """Get cloud self-hosted runners."""
 
     @abc.abstractmethod
+    # Abstract methods do not have a return value, ignore the docstring error DCO031
     def delete_vms(self, instance_ids: Sequence[InstanceID]) -> list[InstanceID]:
         """Delete cloud VM instances.
 
@@ -269,9 +243,10 @@ class CloudRunnerManager(abc.ABC):
 
         Returns:
             The deleted instance IDs.
-        """
+        """  # noqa: DCO031
 
     @abc.abstractmethod
+    # Abstract methods do not have a return value, ignore the docstring error DCO031
     def extract_metrics(self, instance_ids: Sequence[InstanceID]) -> list[RunnerMetrics]:
         """Extract metrics from cloud VMs.
 
@@ -280,7 +255,7 @@ class CloudRunnerManager(abc.ABC):
 
         Returns:
             The fetched runner metrics.
-        """
+        """  # noqa: DCO031
 
     @abc.abstractmethod
     def cleanup(self) -> None:
